Move plan type definitions to definitions.js
